requirements.txt
mido==1.2.9
py-midi==2.0.1
pyserial==3.4
python-rtmidi==1.3.0